Tests/LibWeb/Text/input/css/style-invalidation/sheet-add-keyframes-only-shadow-host.html

<!DOCTYPE html>
<script src="../../include.js"></script>
<div id="host"></div>
<script>
    function settleAndReset(triggerElement) {
        getComputedStyle(triggerElement).animationName;
        getComputedStyle(triggerElement).animationName;
        internals.resetStyleInvalidationCounters();
    }

    function addBystanders(parent, count) {
        for (let i = 0; i < count; ++i) {
            const bystander = document.createElement("div");
            bystander.textContent = `bystander ${i}`;
            parent.appendChild(bystander);
        }
    }

    function dumpCounters(label) {
        const c = internals.getStyleInvalidationCounters();
        println(`${label}: fullStyleInvalidations=${c.fullStyleInvalidations}, styleInvalidations=${c.styleInvalidations}, elementStyleRecomputations=${c.elementStyleRecomputations}, elementStyleNoopRecomputations=${c.elementStyleNoopRecomputations}`);
    }

    test(() => {
        const host = document.getElementById("host");
        const shadowRoot = host.attachShadow({ mode: "open" });

        const hostStyle = document.createElement("style");
        hostStyle.textContent = ":host { animation-name: fade; animation-duration: 1s; }";
        shadowRoot.appendChild(hostStyle);

        // Bystanders inside the shadow subtree highlight the work that a
        // shadow-scope-broad invalidation does even though only the host
        // (which lives outside the shadow subtree, but matches via :host)
        // actually depends on the new keyframes.
        addBystanders(shadowRoot, 25);
        settleAndReset(host);

        // Adding a sheet whose only effect is declaring @keyframes fade should
        // dirty just the host, whose computed animation-name is "fade" via
        // the earlier :host rule.
        const keyframesStyle = document.createElement("style");
        keyframesStyle.textContent = "@keyframes fade { from { opacity: 0.25; } to { opacity: 0.25; } }";
        shadowRoot.appendChild(keyframesStyle);
        getComputedStyle(host).animationName;

        dumpCounters("@keyframes only sheet add (host animation-name via :host)");
    });
</script>
