Tests/LibWeb/Text/expected/css/style-invalidation/sheet-add-keyframes-only-shadow-host.txt

@keyframes only sheet add (host animation-name via :host): fullStyleInvalidations=0, styleInvalidations=3, elementStyleRecomputations=2, elementStyleNoopRecomputations=0
